requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
feedparser>=6.0.11
httpx>=0.27.0
python-dateutil>=2.9.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
atproto>=0.0.55
